requirements.txt
flask==2.0.1
werkzeug==2.0.3
pymongo==4.6.1
python-dotenv==1.0.0
anthropic>=0.7.0 